requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo == 4.13.2
scikit-learn
dill
pyaml
#-e .

